Split BasicJoinBenchmark up.
